turan_talim_bot/requirements.txt
aiogram>=3.2.0
aiohttp>=3.8.5
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncpg>=0.28.0
